+++ pkg/providers/vsphere/vmprovider_vm_test.go
@@ -87,8 +87,8 @@ func vmTests() {
 		)
 
 		BeforeEach(func() {
-			vmClass = builder.DummyVirtualMachineClassA2()
-			vm = builder.DummyBasicVirtualMachineA2("test-vm", "")
+			vmClass = builder.DummyVirtualMachineClassGenName()
+			vm = builder.DummyBasicVirtualMachine("test-vm", "")
 
 			// Reduce diff from old tests: by default don't create an NIC.
 			if vm.Spec.Network == nil {
@@ -116,7 +116,7 @@ func vmTests() {
 				vsphere.SkipVMImageCLProviderCheck = true
 
 				// Use the default VM created by vcsim as the source.
-				clusterVMImage = builder.DummyClusterVirtualMachineImageA2("DC0_C0_RP0_VM0")
+				clusterVMImage = builder.DummyClusterVirtualMachineImage("DC0_C0_RP0_VM0")
 				Expect(ctx.Client.Create(ctx, clusterVMImage)).To(Succeed())
 				conditions.MarkTrue(clusterVMImage, vmopv1.ReadyConditionType)
 				Expect(ctx.Client.Status().Update(ctx, clusterVMImage)).To(Succeed())
