+++ webhooks/virtualmachineclass/mutation/virtualmachineclass_mutator_intg_test.go
@@ -38,7 +38,7 @@ func newIntgMutatingWebhookContext() *intgMutatingWebhookContext {
 		IntegrationTestContext: *suite.NewIntegrationTestContext(),
 	}
 
-	ctx.vmClass = builder.DummyVirtualMachineClassA2()
+	ctx.vmClass = builder.DummyVirtualMachineClassGenName()
 	ctx.vmClass.Namespace = ctx.Namespace
 
 	return ctx
