+++ webhooks/virtualmachineclass/mutation/virtualmachineclass_mutator_unit_test.go
@@ -36,7 +36,7 @@ type unitMutationWebhookContext struct {
 }
 
 func newUnitTestContextForMutatingWebhook() *unitMutationWebhookContext {
-	vmClass := builder.DummyVirtualMachineClassA2()
+	vmClass := builder.DummyVirtualMachineClassGenName()
 	obj, err := builder.ToUnstructured(vmClass)
 	Expect(err).ToNot(HaveOccurred())
