+++ webhooks/virtualmachineclass/validation/virtualmachineclass_validator_intg_test.go
@@ -50,7 +50,7 @@ func newIntgValidatingWebhookContext() *intgValidatingWebhookContext {
 		IntegrationTestContext: *suite.NewIntegrationTestContext(),
 	}
 
-	ctx.vmClass = builder.DummyVirtualMachineClassA2()
+	ctx.vmClass = builder.DummyVirtualMachineClassGenName()
 	ctx.vmClass.Namespace = ctx.Namespace
 
 	return ctx
