+++ webhooks/virtualmachineservice/validation/virtualmachineservice_validator_intg_test.go
@@ -58,7 +58,7 @@ func newIntgValidatingWebhookContext() *intgValidatingWebhookContext {
 		IntegrationTestContext: *suite.NewIntegrationTestContext(),
 	}
 
-	ctx.vmService = builder.DummyVirtualMachineServiceA2()
+	ctx.vmService = builder.DummyVirtualMachineService()
 	ctx.vmService.Namespace = ctx.Namespace
 
 	return ctx
